added SearchForm component for SimpleList
